Documentation for Wahoo.

Wahoo.track

Wahoo.jl

Wahoo is a package designed for tracking the movement of marine animals using probabilistic models

Installation

] add git@github.com:scheidan/Wahoo.jl.git

Usage

The following example uses example data that come with Wahoo.jl.

using Wahoo

import GeoArrays
using DelimitedFiles: readdlm

# This is optional. If *both* packages are imported,
# some computations use the GPU. Otherwise, the CPU is used for everything.
import CUDA
import cuDNN

1) Bathymetry

Load the bathymetry map which provides depth information for each grid cell.

pathdata = joinpath(pkgdir(Wahoo), "example_data")
bathymetry_map = GeoArrays.read(joinpath(pathdata, "bathymetry_200m.tif"))
GeoArrays.bbox(bathymetry_map)

2) Depth observations

Read the depth measurements and define the exponential likelihood model. This model implies the fish is more likely close to the seabed.

# read depth signals
depth_signals = readdlm(joinpath(pathdata, "depth_observations.csv"), ',', header=true)[1][:,2]

# define likelihood
function p_obs_depth_exponential(signals, t, waterdepth, dist; scale=30f0)
    signal = signals[t]
    if signal > waterdepth         # water is too shallow
        return zero(waterdepth)
    else
        # exponential
        Z = 1 - exp(-waterdepth/scale) # normalisation due to truncation
        exp(-(waterdepth - signal)/scale)/(scale * Z)
    end
end

3) Acoustic observations

Load acoustic detection data and specify the probability model for acoustic signals given the distance.

# read acoustic signals
acoustic_signals = readdlm(joinpath(pathdata, "acoustic_observations.csv"), ',', header=true)[1][:,2:3]
acoustic_signals = Int.(acoustic_signals')
acoustic_obs = [acoustic_signals[1,:], acoustic_signals[2,:]]

# read sensor positions
moorings = readdlm(joinpath(pathdata, "acoustic_moorings.csv"), ',', header=true)[1]
acoustic_pos = tuple.(moorings[:,2], moorings[:,3])

# define likelihood
function p_obs_acoustic(signals, t::Int, depth::Number, distance::Number)
    Wahoo.p_acoustic_sigmoid(signals[t], depth, distance)
end

4) Define parameters

We define initial distribution of the fish location and configures the model parameters such as time steps, movement capabilities of the fish, and spatial resolution of the bathymetry.

# initial values: Matrix{Float64}
p0 = zeros(size(bathymetry_map)[1], size(bathymetry_map)[2])
idx = GeoArrays.indices(bathymetry_map, (709757.111649658, 6.26772603565296e6)) # last known location of the fish
p0[idx] = 1

tsave = 1:2:720             # time steps to save
movement_std = 100          # standard deviation of the fish movement for one time step [m]
spatial_resolution = 200    # spatial resolution [m]

5) Run inference

Finally, we run the model inference combine all observations and assumptions.

res = track(pos_init = p0, bathymetry = bathymetry_map,
            tsave = tsave,
            spatial_resolution = spatial_resolution,
            movement_std = movement_std,
            observations = [depth_signals, acoustic_obs...],
            observation_models = [p_obs_depth_exponential, acoustic_obs_models...],
            sensor_positions = [nothing, acoustic_pos...],
            n_trajectories = 2)

# Resulting probabilities
# Array{Float32, 4}: Ny × Nx × 1 × time
res.pos_smoother     # Prob(s_t | y_{1...T})
res.pos_filter       # Prob(s_t | y_{1...t}), only if `save_filter = true` was used
res.residence_dist   # 1/T Σ Prob(s_t | y_{1...T})
res.trajectories     # Vector of trajectories sampled from Prob(s_{1...T} | y_{1...T})
res.log_p            # Prob(y_t)
res.tsave            # time points

Defining observation models

Using GPU

todo

API

Wahoo.track — Function

Tracks the location of the fish

track(;pos_init::Matrix,
       tsave::AbstractVector = 1:100,
       bathymetry::GeoArrays.GeoArray,
       observations::Vector,
       observation_models::Vector{Function},
       sensor_positions::Vector,
       spatial_resolution,
       movement_std,
       save_filter::Bool = false,
       n_trajectories::Int = 0,
       show_progressbar::Bool = !is_logging(stderr),
       precision = Float32)

Infers the location of the animal based on a diffusion model and smoothing.

Keyword Arguments

pos_init::Matrix: Initial probability distribution of the fish position

tsave::AbstractVector: Time steps at which the probability map is saved.

bathymetry: Bathymetric data as GeoArray

spatial_resolution: the spatial resolution [m] of the bathymetry.

movement_std: Standard deviation of the fish movement within one time step [m]

observations: Vector holding all observations. Each element contains the observation of a separate sensor.

observation_models::Vector{Function}: Vector containing the observation model for each sensor.

sensor_positions: Vector of tuples of coordinates or nothing, i.e. Vector{Union{Nothing, Tuple{Real, Real}}}

save_filter: if true the probabilities from the filter run are returned.

n_trajectories=0: Number of trajectories to sample

show_progressbar = !is_logging(stderr): defaults to true for interactive use.

precision = Float32: numerical floating point type used for computations

Note, the elements of the vectors observations, observation_models, and sensor_positions must be sorted in the same way, i.e. the elements at the same position in the Vectors refer to the same sensor.

Return

A named tuple with the following elements:

pos_smoother: Smoothed probability distribution of the fish positions for all timesteps in tsave.

residence_dist: Residence distribution.

trajectories: Sampled trajectories if n_trajectories > 0, otherwise nothing.

log_p: Log probability of the observations.

tsave: Vector of time steps at which the results are saved.

pos_filter: Filtered probability distribution of the fish positions, included if save_filter = true.

References

The filter and smoother implementation is based on:

Thygesen, Uffe Høgsbro, Martin Wæver Pedersen, and Henrik Madsen. 2009. “Geolocating Fish Using Hidden Markov Models and Data Storage Tags.” In Tagging and Tracking of Marine Animals with Electronic Devices, 277–93. Dordrecht: Springer Netherlands. https://doi.org/10.1007/978-1-4020-9640-2_17.