Wahoo

Documentation for Wahoo.

Wahoo.divzero

Wahoo.make_kernel

Wahoo.time2index

Wahoo.track

Wahoo.divzero — Method

Define division by zero equal to zero.

Wahoo.make_kernel — Method

Compute convolution kernel and the required number of hope per time steps

D: diffusion coefficient h: spatial resolution

Wahoo.time2index — Method

Maps time step to the index of pos_filter and pos_smoother

Wahoo.track — Method

Tracks the location of the fish

track(p_init, bathymetry; depth_obs, acoustic_obs, acoustic_pos, tsave, D, h, smoothing, use_gpu)

Uses forward filtering based on a diffusion model and optionally smoothing.

Arguments

p_init::Matrix: Initial probability distribution of the fish positions

bathymetry: Bathymetric data of the environment

depth_obs: A vector of observed depths at each time step

acoustic_obs: A matrix of acoustic observations

acoustic_pos: Vector of tuples containing the positions of the acoustic receivers

tsave::AbstractVector: Time steps at which to save the probability map

D: Diffusion coefficient, i.e. variance for one time step movement [m^2]

h: spatial resolution [m]

smoothing: Boolean flag to enable smoothing