Wahoo

Documentation for Wahoo.

Wahoo.divzero

Wahoo.make_kernel

Wahoo.time2index

Wahoo.track

Wahoo.divzero — Method

Define division by zero as equal to zero.

Wahoo.make_kernel — Method

Compute convolution kernel and the required number of hope per time steps

D: diffusion coefficient h: spatial resolution

Wahoo.time2index — Method

Maps time step to the index of pos_filter and pos_smoother

Wahoo.track — Method

Tracks the location of the fish

track(;pos_init, bathymetry, observations,  observation_models, sensor_positions, tsave, h, D, smoothing=true)

Uses forward filtering based on a diffusion model and optionally smoothing.

Arguments

pos_init::Matrix: Initial probability distribution of the fish positions

bathymetry: Bathymetric data of the environment

observations: Vector of all observations

sensor_positions: Vector of tuples or nothing containing the positions of the receivers

tsave::AbstractVector: Time steps at which to save the probability map

h: spatial resolution [m]

D: Diffusion coefficient, i.e. variance for one time step movement [m^2]

smoothing: Boolean flag to enable smoothing

show_progressbar = !is_logging(stderr): defaults to true for interactive use.

precision = Float32: numerical floating point type used for computations